typeinfo数据落盘
# message auto-generated for no-merge-commit merge:
merge ctrlc0925 into master

typeinfo数据落盘

Created-by: msprofDeveloper

Match-id-b4640a6df0f24b6ba57d51a7d58125cb593d6630

+++ collector/dvvp/transport/hash_data.cpp
@@ -285,7 +285,7 @@ void HashData::SaveNewHashData(bool isLastChunk)
     }
     // combined hash map data
     std::string saveHashData;
-    int currentHashSize = hashVector_.size();
+    size_t currentHashSize = hashVector_.size();
     for (size_t i = readIndex_; i < currentHashSize; i++) {
         saveHashData.append(std::to_string(hashVector_[i].first) + HASH_DIC_DELIMITER + hashVector_[i].second + "\n");
     }
